minor fixes; libnxlp and logwatch ported to NetWare

+++ src/libnxlp/rule.cpp
@@ -148,7 +148,7 @@ void LogParserRule::ExpandMacros(const char *regexp, String &out)
 			// Check for escape
 			if ((curr != regexp) && (*(curr - 1) == '\\'))
 			{
-				out.AddString(prev, curr - prev - 1);
+				out.AddString(prev, (DWORD)(curr - prev - 1));
 				out += _T("@");
 			}
 			else
@@ -158,7 +158,7 @@ void LogParserRule::ExpandMacros(const char *regexp, String &out)
 				{
 					int i;
 
-					out.AddString(prev, curr - prev);
+					out.AddString(prev, (DWORD)(curr - prev));
 					curr += 2;
 					for(i = 0; (*curr != 0) && (*curr != '}'); i++)
 						name[i] = *curr++;
@@ -167,11 +167,11 @@ void LogParserRule::ExpandMacros(const char *regexp, String &out)
 				}
 				else
 				{
-					out.AddString(prev, curr - prev + 1);
+					out.AddString(prev, (DWORD)(curr - prev + 1));
 				}
 			}
 			prev = curr + 1;
 		}
 	}
-	out.AddString(prev, curr - prev);
+	out.AddString(prev, (DWORD)(curr - prev));
 }
